+++ src/main/java/name/maxdeliso/micron/peer/Peer.java
@@ -4,15 +4,15 @@ import lombok.Value;
 import net.jcip.annotations.ThreadSafe;
 
 import java.nio.channels.SocketChannel;
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicInteger;
 
 @ThreadSafe
 @Value
 public final class Peer implements PositionTracker {
 
-  private final long index;
+  private final int index;
 
-  private final AtomicLong position;
+  private final AtomicInteger position;
 
   private final SocketChannel socketChannel;
 
@@ -22,19 +22,19 @@ public final class Peer implements PositionTracker {
    * @param index         the numeric index of the peer.
    * @param socketChannel a selectable channel with which to communicate with them.
    */
-  public Peer(final long index, final SocketChannel socketChannel) {
+  public Peer(final int index, final SocketChannel socketChannel) {
     this.index = index;
-    this.position = new AtomicLong(0);
+    this.position = new AtomicInteger(0);
     this.socketChannel = socketChannel;
   }
 
   @Override
-  public long advancePosition() {
+  public int advancePosition() {
     return this.position.incrementAndGet();
   }
 
   @Override
-  public long advancePosition(final long delta) {
+  public int advancePosition(final int delta) {
     return this.position.addAndGet(delta);
   }
 
@@ -44,7 +44,7 @@ public final class Peer implements PositionTracker {
   }
 
   @Override
-  public long getPosition() {
+  public int getPosition() {
     return this.position.get();
   }
 }
